# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "MIRI API",
  "description": "데이터 기반 상권 분석·업종 추천 서비스 MIRI의 REST API. 로그인 없이 누구나 쓰고, 저장 리포트는 브라우저마다 따로 보관한다(X-Device-Key 헤더). 금액 단위는 원(KRW).",
  "version": "0.2.0"
 },
 "paths": {
...
  "/api/v1/places/search": {
   "get": {
    "tags": [
     "상권 (SB-02)"
    ],
    "summary": "지역·상권 검색 (예: 강남역)",
    "operationId": "search_places_api_v1_places_search_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 1,
       "maxLength": 50,
       "title": "Q"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 20,
       "minimum": 1,
       "default": 8,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/PlaceOut"
         },
         "title": "Response Search Places Api V1 Places Search Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/areas/candidates": {
   "get": {
    "tags": [
     "상권 (SB-02)"
    ],
    "summary": "반경 내 후보 상권",
    "operationId": "area_candidates_api_v1_areas_candidates_get",
    "parameters": [
     {
      "name": "lat",
      "in": "query",
      "required": true,
      "schema": {
       "type": "number",
       "maximum": 39,
       "minimum": 33,
       "title": "Lat"
      }
     },
     {
      "name": "lng",
      "in": "query",
      "required": true,
      "schema": {
       "type": "number",
       "maximum": 132,
       "minimum": 124,
       "title": "Lng"
      }
     },
     {
      "name": "radius_m",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 5000,
       "minimum": 100,
       "default": 500,
       "title": "Radius M"
      }
     },
     {
      "name": "place_name",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 100
        },
        {
         "type": "null"
        }
       ],
       "title": "Place Name"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CandidatesOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/areas/compare": {
   "post": {
    "tags": [
     "상권 (SB-02)"
    ],
    "summary": "상권 비교 (UC-02) — SB-02 '후보 비교' 시트",
    "operationId": "area_compare_api_v1_areas_compare_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CompareIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CompareOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/areas/{area_code}": {
   "get": {
    "tags": [
     "상권 (SB-02)"
    ],
    "summary": "상권 기본 지표",
    "operationId": "area_detail_api_v1_areas__area_code__get",
    "parameters": [
     {
      "name": "area_code",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Area Code"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AreaCardOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/analyses": {
   "post": {
    "tags": [
     "분석 (SB-03~08)"
    ],
    "summary": "분석 실행 → 업종 TOP 5 (UC-01)",
    "operationId": "create_analysis_api_v1_analyses_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AnalysisIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AnalysisOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/analyses/{analysis_id}": {
   "get": {
    "tags": [
     "분석 (SB-03~08)"
    ],
    "summary": "분석 결과 다시 보기",
    "operationId": "get_analysis_api_v1_analyses__analysis_id__get",
    "parameters": [
     {
      "name": "analysis_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Analysis Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AnalysisOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/analyses/{analysis_id}/risk/{industry_code}": {
   "get": {
    "tags": [
     "분석 (SB-03~08)"
    ],
    "summary": "폐업 위험도 상세 (UC-03)",
    "operationId": "get_risk_api_v1_analyses__analysis_id__risk__industry_code__get",
    "parameters": [
     {
      "name": "analysis_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Analysis Id"
      }
     },
     {
      "name": "industry_code",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Industry Code"
      }
     },
     {
      "name": "be",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer",
         "maximum": 9223372036854775807,
         "minimum": 1
        },
        {
         "type": "null"
        }
       ],
       "description": "리포트에서 열 때 그 리포트에 반영한 손익분기 id — 인근 지역이 비용을 감당하는지(alternatives[].affordable)를 그 비용으로 판단",
       "title": "Be"
      },
      "description": "리포트에서 열 때 그 리포트에 반영한 손익분기 id — 인근 지역이 비용을 감당하는지(alternatives[].affordable)를 그 비용으로 판단"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RiskOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/analyses/{analysis_id}/breakeven": {
   "post": {
    "tags": [
     "분석 (SB-03~08)"
    ],
    "summary": "손익분기점 계산 (UC-04)",
    "operationId": "post_breakeven_api_v1_analyses__analysis_id__breakeven_post",
    "parameters": [
     {
      "name": "analysis_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Analysis Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BreakEvenIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BreakEvenOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/analyses/{analysis_id}/breakeven/{break_even_id}": {
   "get": {
    "tags": [
     "분석 (SB-03~08)"
    ],
    "summary": "저장된 손익분기 결과",
    "operationId": "get_breakeven_api_v1_analyses__analysis_id__breakeven__break_even_id__get",
    "parameters": [
     {
      "name": "analysis_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Analysis Id"
      }
     },
     {
      "name": "break_even_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "maximum": 9223372036854775807,
       "minimum": 1,
       "title": "Break Even Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BreakEvenOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/analyses/{analysis_id}/reports": {
   "post": {
    "tags": [
     "분석 (SB-03~08)"
    ],
    "summary": "액션 리포트 생성 (UC-05)",
    "operationId": "post_report_api_v1_analyses__analysis_id__reports_post",
    "parameters": [
     {
      "name": "analysis_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Analysis Id"
      }
     },
     {
      "name": "X-Device-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "이 브라우저의 저장 키(없으면 저장 여부를 알려 주지 않음)",
       "title": "X-Device-Key"
      },
      "description": "이 브라우저의 저장 키(없으면 저장 여부를 알려 주지 않음)"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReportIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReportOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/refresh": {
   "post": {
    "tags": [
     "관리자 (UC-07)"
    ],
    "summary": "원천 재적재 + 특징·모델 배치 + 엔진 재로딩",
    "operationId": "refresh_api_v1_admin_refresh_post",
    "parameters": [
     {
      "name": "skip_ingest",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Skip Ingest"
      }
     },
     {
      "name": "X-Admin-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "서버에 지정한 MIRI_ADMIN_KEY",
       "title": "X-Admin-Key"
      },
      "description": "서버에 지정한 MIRI_ADMIN_KEY"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/ingestion-logs": {
   "get": {
    "tags": [
     "관리자 (UC-07)"
    ],
    "summary": "수집 로그",
    "operationId": "ingestion_logs_api_v1_admin_ingestion_logs_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "X-Admin-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "서버에 지정한 MIRI_ADMIN_KEY",
       "title": "X-Admin-Key"
      },
      "description": "서버에 지정한 MIRI_ADMIN_KEY"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "AnalysisIn": {
    "properties": {
     "area_code": {
      "type": "string",
      "maxLength": 20,
      "minLength": 1,
      "title": "Area Code",
      "description": "분석 대상 지역 코드(후보 카드에서 선택)"
     },
     "place_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100
       },
       {
        "type": "null"
       }
      ],
      "title": "Place Name"
     },
     "lat": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 39.0,
        "minimum": 33.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Lat"
     },
     "lng": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 132.0,
        "minimum": 124.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Lng"
     },
     "radius_m": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 5000.0,
        "minimum": 100.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Radius M"
     },
     "candidate_area_codes": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 10,
      "title": "Candidate Area Codes"
     },
     "user_type": {
      "type": "string",
      "enum": [
       "PRE_FOUNDER",
       "OWNER",
       "CONSULTANT"
      ],
      "title": "User Type",
      "default": "PRE_FOUNDER"
     },
     "budget": {
      "type": "integer",
      "maximum": 100000000000.0,
      "minimum": 0.0,
      "title": "Budget",
      "description": "총 창업 예산(원)"
     },
     "monthly_rent_limit": {
      "type": "integer",
      "maximum": 100000000000.0,
      "minimum": 0.0,
      "title": "Monthly Rent Limit",
      "description": "월 임대료(원). 구하려는 점포 월세 — 아직 정하지 않았으면 감당할 수 있는 최대치"
     },
     "labor_cost": {
      "type": "integer",
      "maximum": 100000000000.0,
      "minimum": 0.0,
      "title": "Labor Cost",
      "description": "예상 인건비(월, 원)"
     },
     "initial_investment": {
      "type": "integer",
      "maximum": 100000000000.0,
      "minimum": 0.0,
      "title": "Initial Investment",
      "description": "인테리어·설비 등 초기 투자비(보증금 제외)",
      "default": 0
     },
     "other_fixed": {
      "type": "integer",
      "maximum": 100000000000.0,
      "minimum": 0.0,
      "title": "Other Fixed",
      "description": "관리비·공과금 등 기타 월 고정비",
      "default": 0
     },
     "loan_amount": {
      "type": "integer",
      "maximum": 100000000000.0,
      "minimum": 0.0,
      "title": "Loan Amount",
      "default": 0
     },
     "loan_rate_annual": {
      "type": "number",
      "maximum": 0.3,
      "minimum": 0.0,
      "title": "Loan Rate Annual",
      "description": "대출 연이율(0.05 = 5%)",
      "default": 0.0
     },
     "owner_salary": {
      "type": "integer",
      "maximum": 100000000000.0,
      "minimum": 0.0,
      "title": "Owner Salary",
      "description": "목표 월수입(대표자, 선택). 넣으면 손익분기에 포함",
      "default": 0
     },
     "business_goal": {
      "type": "string",
      "enum": [
       "기본",
       "안정형",
       "고수익형",
       "저비용형"
      ],
      "title": "Business Goal",
      "default": "기본"
     },
     "licenses": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 20,
      "title": "Licenses",
      "description": "보유 자격(업종 목록의 license 값만)"
     },
     "categories": {
      "items": {
       "type": "string",
       "enum": [
        "외식업",
        "서비스업",
        "소매업"
       ]
      },
      "type": "array",
      "maxItems": 3,
      "title": "Categories"
     },
     "excluded_industries": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 100,
      "title": "Excluded Industries"
     },
     "interests": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 20,
      "title": "Interests",
      "description": "관심 업종 코드(SB-04)"
     },
     "cogs_rates": {
      "additionalProperties": {
       "type": "number"
      },
      "propertyNames": {
       "enum": [
        "외식업",
        "서비스업",
        "소매업"
       ]
      },
      "type": "object",
      "title": "Cogs Rates",
      "description": "분야별 원가율(선택, 0~0.95). 넣은 분야는 비용 적합도(F)·손익분기 기본값에 쓰인다. 외식업은 비우면 평균 40.7%, 서비스업·소매업은 비우면 비용 적합도를 계산하지 않는다"
     }
    },
    "type": "object",
    "required": [
     "area_code",
     "budget",
     "monthly_rent_limit",
     "labor_cost"
    ],
    "title": "AnalysisIn"
   },
   "AnalysisOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "area": {
      "$ref": "#/components/schemas/AnalysisAreaOut"
     },
     "conditions": {
      "additionalProperties": true,
      "type": "object",
      "title": "Conditions"
     },
     "data_quarter": {
      "type": "integer",
      "title": "Data Quarter"
     },
     "data_quarter_label": {
      "type": "string",
      "title": "Data Quarter Label"
     },
     "model_version": {
      "type": "string",
      "title": "Model Version"
     },
     "business_goal": {
      "type": "string",
      "title": "Business Goal"
     },
     "weights": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Weights"
     },
     "f_applied": {
      "type": "boolean",
      "title": "F Applied"
     },
     "f_note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "F Note"
     },
     "n_candidates": {
      "type": "integer",
      "title": "N Candidates"
     },
     "n_low_score": {
      "type": "integer",
      "title": "N Low Score",
      "description": "조건은 통과했지만 적합도가 20점 미만이라 TOP에 넣지 않은 업종 수",
      "default": 0
     },
     "empty_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Empty Reason",
      "description": "추천 TOP이 0개일 때 가장 큰 원인(예: 월 고정비가 모든 업종의 평균 매출보다 큼)"
     },
     "top": {
      "items": {
       "$ref": "#/components/schemas/ItemOut"
      },
      "type": "array",
      "title": "Top"
     },
     "interests": {
      "items": {
       "$ref": "#/components/schemas/ItemOut"
      },
      "type": "array",
      "title": "Interests"
     },
     "cautions": {
      "items": {
       "$ref": "#/components/schemas/ItemOut"
      },
      "type": "array",
      "title": "Cautions"
     },
     "missing_interests": {
      "items": {
       "$ref": "#/components/schemas/MissingInterestOut"
      },
      "type": "array",
      "title": "Missing Interests",
      "description": "이 지역에 점포가 없어 점수를 못 낸 관심 업종"
     },
     "notices": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Notices"
     }
    },
    "type": "object",
    "required": [
     "id",
     "created_at",
     "area",
     "conditions",
     "data_quarter",
     "data_quarter_label",
     "model_version",
     "business_goal",
     "weights",
     "f_applied",
     "n_candidates",
     "top",
     "interests",
     "cautions",
     "notices"
    ],
    "title": "AnalysisOut"
   },
   "AreaCardOut": {
    "properties": {
     "area_code": {
      "type": "string",
      "title": "Area Code"
     },
     "area_name": {
      "type": "string",
      "title": "Area Name"
     },
     "level": {
      "type": "string",
      "title": "Level"
     },
     "is_center": {
      "type": "boolean",
      "title": "Is Center",
      "default": false
     },
     "distance_m": {
      "type": "integer",
      "title": "Distance M",
      "default": 0
     },
     "lat": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lat"
     },
     "lng": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lng"
     },
     "metrics": {
      "$ref": "#/components/schemas/AreaMetricsOut"
     }
    },
    "type": "object",
    "required": [
     "area_code",
     "area_name",
     "level",
     "metrics"
    ],
    "title": "AreaCardOut"
   },
...
   "BreakEvenIn": {
    "properties": {
     "industry_code": {
      "type": "string",
      "title": "Industry Code"
     },
     "monthly_rent": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 100000000000.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Monthly Rent",
      "description": "미입력 시 분석 조건의 월 임대료"
     },
     "labor_cost": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 100000000000.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Labor Cost"
     },
     "other_fixed": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 100000000000.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Other Fixed"
     },
     "initial_investment": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 100000000000.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Initial Investment"
     },
     "loan_amount": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 100000000000.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Loan Amount"
     },
     "loan_rate_annual": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 0.3,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Loan Rate Annual"
     },
     "owner_salary": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 100000000000.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Owner Salary"
     },
     "cogs_rate": {
      "anyOf": [
       {
        "type": "number",
        "exclusiveMaximum": 1.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Cogs Rate",
      "description": "원가율. 비우면 분석 조건의 분야별 원가율(cogs_rates) → 업종 대분류 평균(외식업만). 둘 다 없으면 COGS_REQUIRED"
     },
     "other_variable_rate": {
      "anyOf": [
       {
        "type": "number",
        "exclusiveMaximum": 1.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Other Variable Rate",
      "description": "배달·플랫폼 수수료 등"
     },
     "avg_ticket": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 10000000.0,
        "minimum": 100.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Avg Ticket",
      "description": "미입력 시 이 지역 동일 업종 실제 객단가, 그것도 없으면 서울 같은 업종 평균 객단가(100원 이상)"
     }
    },
    "type": "object",
    "required": [
     "industry_code"
    ],
    "title": "BreakEvenIn"
   },
...
   "CandidatesOut": {
    "properties": {
     "lat": {
      "type": "number",
      "title": "Lat"
     },
     "lng": {
      "type": "number",
      "title": "Lng"
     },
     "radius_m": {
      "type": "integer",
      "title": "Radius M"
     },
     "place_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Place Name"
     },
     "analysis_level": {
      "type": "string",
      "title": "Analysis Level"
     },
     "notice": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notice"
     },
     "candidates": {
      "items": {
       "$ref": "#/components/schemas/AreaCardOut"
      },
      "type": "array",
      "title": "Candidates"
     }
    },
    "type": "object",
    "required": [
     "lat",
     "lng",
     "radius_m",
     "analysis_level",
     "candidates"
    ],
    "title": "CandidatesOut"
   },
...
   "CompareIn": {
    "properties": {
     "area_codes": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 4,
      "minItems": 1,
      "title": "Area Codes"
     },
     "industry_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Industry Code"
     }
    },
    "type": "object",
    "required": [
     "area_codes"
    ],
    "title": "CompareIn"
   },
...
   "CompareOut": {
    "properties": {
     "industry_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Industry Code"
     },
     "industry_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Industry Name"
     },
     "data_quarter": {
      "type": "integer",
      "title": "Data Quarter"
     },
     "areas": {
      "items": {
       "$ref": "#/components/schemas/CompareAreaOut"
      },
      "type": "array",
      "title": "Areas"
     }
    },
    "type": "object",
    "required": [
     "data_quarter",
     "areas"
    ],
    "title": "CompareOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "PlaceOut": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "lat": {
      "type": "number",
      "title": "Lat"
     },
     "lng": {
      "type": "number",
      "title": "Lng"
     },
     "area_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Area Code"
     },
     "area_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Area Name"
     },
     "address": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "source": {
      "type": "string",
      "enum": [
       "local",
       "kakao"
      ],
      "title": "Source",
      "default": "local"
     }
    },
    "type": "object",
    "required": [
     "name",
     "kind",
     "lat",
     "lng"
    ],
    "title": "PlaceOut"
   },
...
   "ReportIn": {
    "properties": {
     "industry_code": {
      "type": "string",
      "title": "Industry Code"
     },
     "break_even_id": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 9.223372036854776e+18,
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Break Even Id"
     }
    },
    "type": "object",
    "required": [
     "industry_code"
    ],
    "title": "ReportIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}